+++ MiBazarProyecto/src/main/java/com/nicodev/MiBazarProyecto/model/Producto.java
@@ -17,12 +17,12 @@ public class Producto {
     private String nombre;
     private String marca;
     public Double costo;
-    private Double cantidad_disponible;
+    private Integer cantidad_disponible; // Integer = Wrapper, acepta Null.
 
     public Producto() {
     }
 
-    public Producto(Long codigo_producto, String nombre, String marca, Double costo, Double cantidad_disponible) {
+    public Producto(Long codigo_producto, String nombre, String marca, Double costo, Integer cantidad_disponible) {
         this.codigo_producto = codigo_producto;
         this.nombre = nombre;
         this.marca = marca;
