+++ MiBazarProyecto/src/main/java/com/nicodev/MiBazarProyecto/service/IProductoService.java
@@ -19,7 +19,7 @@ public interface IProductoService {
     public void deleteProducto(Long codigo_producto);
 
     // PUT/EDIT un producto (por ID)
-    public void editProducto(Long codigo_producto, String nombre_nue, String marca_nue, Double costo_nue, Double cant_dis_nue);
+    public void editProducto(Long codigo_producto, String nombre_nue, String marca_nue, Double costo_nue, Integer cant_dis_nue);
 
     // 4. Obtener todos los productos cuya cantidad disponible sea menor a 5
     public List<Producto> traerProductoEscaso();
